Experiment in trying to handle stackframes for vscode
For some roms this seems to work pretty well. However, a lot of cases were observed in, e.g., A Link to the Past or F-Zero where the stack management gets a bit bizarre. ALttP shifts the stack pointer around in some cases, making it harder to track what stored stack frames are still valid or pertinent. F-Zero has some behaviour sometimes where it doesn't do JSL to jump, but rather manually manually pushes the bank and addr to return to, then does a JMP, which does an RTL by grabbing the _manually_ added IP to the stack.